+++ datanode/gateway/graphql/generated.go
@@ -1558,6 +1558,7 @@ type ComplexityRoot struct {
 	}
 
 	UpdateAsset struct {
+		AssetId func(childComplexity int) int
 		Quantum func(childComplexity int) int
 		Source  func(childComplexity int) int
 	}
@@ -8914,6 +8915,13 @@ func (e *executableSchema) Complexity(typeName, field string, childComplexity in
 
 		return e.complexity.TransferResponses.Responses(childComplexity), true
 
+	case "UpdateAsset.assetId":
+		if e.complexity.UpdateAsset.AssetId == nil {
+			break
+		}
+
+		return e.complexity.UpdateAsset.AssetId(childComplexity), true
+
 	case "UpdateAsset.quantum":
 		if e.complexity.UpdateAsset.Quantum == nil {
 			break
@@ -12602,6 +12610,9 @@ type UpdateAsset {
 
   "The source of the updated asset"
   source: UpdateAssetSource!
+
+  "The asset to update"
+  assetId: ID!
 }
 
 """
@@ -47463,6 +47474,41 @@ func (ec *executionContext) _UpdateAsset_source(ctx context.Context, field graph
 	return ec.marshalNUpdateAssetSource2codeᚗvegaprotocolᚗioᚋvegaᚋdatanodeᚋgatewayᚋgraphqlᚐUpdateAssetSource(ctx, field.Selections, res)
 }
 
+func (ec *executionContext) _UpdateAsset_assetId(ctx context.Context, field graphql.CollectedField, obj *vega.UpdateAsset) (ret graphql.Marshaler) {
+	defer func() {
+		if r := recover(); r != nil {
+			ec.Error(ctx, ec.Recover(ctx, r))
+			ret = graphql.Null
+		}
+	}()
+	fc := &graphql.FieldContext{
+		Object:     "UpdateAsset",
+		Field:      field,
+		Args:       nil,
+		IsMethod:   false,
+		IsResolver: false,
+	}
+
+	ctx = graphql.WithFieldContext(ctx, fc)
+	resTmp, err := ec.ResolverMiddleware(ctx, func(rctx context.Context) (interface{}, error) {
+		ctx = rctx // use context from middleware stack in children
+		return obj.AssetId, nil
+	})
+	if err != nil {
+		ec.Error(ctx, err)
+		return graphql.Null
+	}
+	if resTmp == nil {
+		if !graphql.HasFieldError(ctx, fc) {
+			ec.Errorf(ctx, "must not be null")
+		}
+		return graphql.Null
+	}
+	res := resTmp.(string)
+	fc.Result = res
+	return ec.marshalNID2string(ctx, field.Selections, res)
+}
+
 func (ec *executionContext) _UpdateERC20_lifetimeLimit(ctx context.Context, field graphql.CollectedField, obj *UpdateErc20) (ret graphql.Marshaler) {
 	defer func() {
 		if r := recover(); r != nil {
@@ -65970,6 +66016,16 @@ func (ec *executionContext) _UpdateAsset(ctx context.Context, sel ast.SelectionS
 				return innerFunc(ctx)
 
 			})
+		case "assetId":
+			innerFunc := func(ctx context.Context) (res graphql.Marshaler) {
+				return ec._UpdateAsset_assetId(ctx, field, obj)
+			}
+
+			out.Values[i] = innerFunc(ctx)
+
+			if out.Values[i] == graphql.Null {
+				atomic.AddUint32(&invalids, 1)
+			}
 		default:
 			panic("unknown field " + strconv.Quote(field.Name))
 		}